carousel slider nearly done, mini image row isnt in sync

+++ lib/classes/product_class.dart
@@ -78,7 +78,7 @@ class Product {
       brandId: json['brand_id'],
       brandName: json['brandName'],
       brandLogoUrl: json['brandImage'],
-      rating: json['product_rating'],
+      rating: json['product_rating'].toDouble(),
       variations: variations,
       availableProperties: availableProperties,
     );
